PR FIXUP - Move definitions to new file
